app/requirements.txt
kafka-python
cassandra-driver
pyspark==3.3.0
pandas
streamlit